Create a condition to avoid users receiving invitations if not allowed
