+++ app/Console/Commands/EnableDefaultNotificationTypes.php
@@ -4,7 +4,6 @@ namespace App\Console\Commands;
 
 use App\Models\NotificationType;
 use App\Models\User;
-use Database\Seeders\NotificationTypeSeederFirstRelease;
 use Illuminate\Console\Command;
 use function Laravel\Prompts\progress;
 
@@ -30,7 +29,7 @@ class EnableDefaultNotificationTypes extends Command
     public function handle(): int
     {
         $users = User::withoutGlobalScopes()->get();
-        $notificationTypes = NotificationType::whereIn('name',NotificationTypeSeederFirstRelease::DEFAULT_NOTIFICATIONS)
+        $notificationTypes = NotificationType::whereIn('name',NotificationType::DEFAULT_NOTIFICATIONS)
             ->get()
             ->pluck('id')
             ->toArray();
