+++ app/Listeners/AfterUserCreated.php
@@ -4,7 +4,6 @@ namespace App\Listeners;
 
 use App\Models\NotificationType;
 use App\Models\User;
-use Database\Seeders\NotificationTypeSeederFirstRelease;
 use Filament\Events\Auth\Registered;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Queue\InteractsWithQueue;
@@ -25,7 +24,7 @@ class AfterUserCreated
     public function handle(Registered $event): void
     {
         $user = User::find($event->getUser()->id);
-        $notificationTypes = NotificationType::whereIn('name',NotificationTypeSeederFirstRelease::DEFAULT_NOTIFICATIONS)
+        $notificationTypes = NotificationType::whereIn('name',NotificationType::DEFAULT_NOTIFICATIONS)
             ->get()
             ->pluck('id')
             ->toArray();
